additional vista admin check: only do if total bytes recvd is less than 100M
git-svn-id: http://svn.vuze.com/public/client@16514 c2c1b7af-1477-405b-86ee-8aefd1fc4dd2

+++ trunk/azureus3/src/com/aelitis/azureus/ui/swt/shells/main/MainWindow.java
@@ -163,7 +163,8 @@ public class MainWindow
 							13).getTimeInMillis()) {
 				// install older than 3016
 				GlobalManager gm = core.getGlobalManager();
-				if (gm != null && gm.getDownloadManagers().size() == 0) {
+				if (gm != null && gm.getDownloadManagers().size() == 0
+						&& gm.getStats().getTotalProtocolBytesReceived() < 1024 * 1024 * 100) {
 					File fileTestWrite = FileUtil.getApplicationFile("testwrite.dll");
 					fileTestWrite.deleteOnExit();
 					try {
